Optimize: Modify the mechanism between the task and change json, and do not redundantly apply the change json content to the task model

+++ packages/watchmen-collector-surface/src/watchmen_collector_surface/task/clean_of_timeout.py
@@ -144,20 +144,21 @@ class CleanOfTimeout:
 		# noinspection PyTypeChecker
 		def clean_change_json_with_task(task: ScheduledTask):
 			for change_json_id in task.changeJsonIds:
-				try:
-					self.change_data_json_service.begin_transaction()
-					change_data_json = self.change_data_json_service.find_by_id(change_json_id)
-					change_data_json.status = Status.FAIL.value
-					change_data_json.result = "timeout"
-					self.change_data_json_history_service.create(change_data_json)
-					# noinspection PyTypeChecker
-					self.change_data_json_service.delete(change_data_json.changeJsonId)
-					self.change_data_json_service.commit_transaction()
-				except Exception as e:
-					self.change_data_json_service.rollback_transaction()
-					raise e
-				finally:
-					self.change_data_json_service.close_transaction()
+				change_data_json = self.change_data_json_service.find_json_by_id(change_json_id)
+				if change_data_json:
+					try:
+						self.change_data_json_service.begin_transaction()
+						change_data_json.status = Status.FAIL.value
+						change_data_json.result = "timeout"
+						self.change_data_json_history_service.create(change_data_json)
+						# noinspection PyTypeChecker
+						self.change_data_json_service.delete(change_data_json.changeJsonId)
+						self.change_data_json_service.commit_transaction()
+					except Exception as e:
+						self.change_data_json_service.rollback_transaction()
+						raise e
+					finally:
+						self.change_data_json_service.close_transaction()
 
 		for task in tasks:
 			clean_change_json_with_task(task)
